pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "test-tecnico-ai-engineer"
version = "0.1.0"
description = "Dataset, evaluation scripts and metrics (Jaccard similarity) for the AI engineer technical test."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
  { name = "SaraCalla" },
]
dependencies = [
    "click>=8.1.8",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pypdf>=4.0.0",
    "datapizza-ai>=0.0.9",
    "datapizza-ai-rerankers-cohere>=0.0.4",
    "loguru>=0.7.0"
]

[tool.setuptools.packages.find]
include = ["src*"]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pytest>=8.0.0",
]
